netdog: Support versioning in `net.toml`
This change adds the ability to handle additional versions of `net.toml`
as the project continues to add features to network configuration.  A
top level `NetConfig` type has been removed in favor of a few traits.
Each new version of config is expected to implement these traits, the
most important of which converts the network configuration into
`WickedInterface` structs which are suitable for serializing directly to
file.

As part of this change, the `net_config` module has been lightly
refactored and split into submodules to make the pieces easier to use
among new versions.
