+++ sources/api/netdog/src/wicked.rs
@@ -4,7 +4,7 @@
 //! The structures in this module are meant to be created from the user-facing structures in the
 //! `net_config` module.  `Default` implementations for WickedInterface exist here as well.
 use crate::interface_name::InterfaceName;
-use crate::net_config::{Dhcp4Config, Dhcp4Options, Dhcp6Config, Dhcp6Options, NetInterface};
+use crate::net_config::{Dhcp4ConfigV1, Dhcp4OptionsV1, Dhcp6ConfigV1, Dhcp6OptionsV1};
 use serde::Serialize;
 use snafu::ResultExt;
 use std::fs;
@@ -16,19 +16,19 @@ const WICKED_FILE_EXT: &str = "xml";
 #[derive(Debug, Serialize, PartialEq)]
 #[serde(rename = "interface")]
 pub(crate) struct WickedInterface {
-    name: InterfaceName,
-    control: WickedControl,
+    pub(crate) name: InterfaceName,
+    pub(crate) control: WickedControl,
     #[serde(skip_serializing_if = "Option::is_none")]
     #[serde(rename = "ipv4:dhcp")]
-    ipv4_dhcp: Option<WickedDhcp4>,
+    pub(crate) ipv4_dhcp: Option<WickedDhcp4>,
     #[serde(skip_serializing_if = "Option::is_none")]
     #[serde(rename = "ipv6:dhcp")]
-    ipv6_dhcp: Option<WickedDhcp6>,
+    pub(crate) ipv6_dhcp: Option<WickedDhcp6>,
 }
 
 #[derive(Debug, Serialize, PartialEq)]
 #[serde(rename_all = "kebab-case")]
-struct WickedControl {
+pub(crate) struct WickedControl {
     #[serde(skip_serializing_if = "Option::is_none")]
     mode: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -122,21 +122,21 @@ struct AddrConfFlags {
     _f: (),
 }
 
-impl From<Dhcp4Config> for WickedDhcp4 {
-    fn from(dhcp4: Dhcp4Config) -> Self {
+impl From<Dhcp4ConfigV1> for WickedDhcp4 {
+    fn from(dhcp4: Dhcp4ConfigV1) -> Self {
         match dhcp4 {
-            Dhcp4Config::DhcpEnabled(b) => WickedDhcp4 {
+            Dhcp4ConfigV1::DhcpEnabled(b) => WickedDhcp4 {
                 enabled: b,
                 _f: (),
                 ..Default::default()
             },
-            Dhcp4Config::WithOptions(o) => WickedDhcp4::from(o),
+            Dhcp4ConfigV1::WithOptions(o) => WickedDhcp4::from(o),
         }
     }
 }
 
-impl From<Dhcp4Options> for WickedDhcp4 {
-    fn from(options: Dhcp4Options) -> Self {
+impl From<Dhcp4OptionsV1> for WickedDhcp4 {
+    fn from(options: Dhcp4OptionsV1) -> Self {
         let mut defer_timeout = None;
         let mut flags = None;
 
@@ -155,21 +155,21 @@ impl From<Dhcp4Options> for WickedDhcp4 {
     }
 }
 
-impl From<Dhcp6Config> for WickedDhcp6 {
-    fn from(dhcp6: Dhcp6Config) -> Self {
+impl From<Dhcp6ConfigV1> for WickedDhcp6 {
+    fn from(dhcp6: Dhcp6ConfigV1) -> Self {
         match dhcp6 {
-            Dhcp6Config::DhcpEnabled(b) => WickedDhcp6 {
+            Dhcp6ConfigV1::DhcpEnabled(b) => WickedDhcp6 {
                 enabled: b,
                 _f: (),
                 ..Default::default()
             },
-            Dhcp6Config::WithOptions(o) => WickedDhcp6::from(o),
+            Dhcp6ConfigV1::WithOptions(o) => WickedDhcp6::from(o),
         }
     }
 }
 
-impl From<Dhcp6Options> for WickedDhcp6 {
-    fn from(options: Dhcp6Options) -> Self {
+impl From<Dhcp6OptionsV1> for WickedDhcp6 {
+    fn from(options: Dhcp6OptionsV1) -> Self {
         let mut defer_timeout = None;
         let mut flags = None;
 
@@ -188,22 +188,7 @@ impl From<Dhcp6Options> for WickedDhcp6 {
 }
 
 impl WickedInterface {
-    /// Create a WickedInterface given a name and configuration
-    pub(crate) fn from_config(name: InterfaceName, config: NetInterface) -> Self {
-        let wicked_dhcp4 = config.dhcp4.map(WickedDhcp4::from);
-        // As additional options are added for IPV6, implement `From` similar to WickedDhcp4
-        let wicked_dhcp6 = config.dhcp6.map(WickedDhcp6::from);
-        WickedInterface {
-            name,
-            control: WickedControl::default(),
-            ipv4_dhcp: wicked_dhcp4,
-            ipv6_dhcp: wicked_dhcp6,
-        }
-    }
-
     /// Serialize the interface's configuration file
-    // Consume `self` to enforce that changes aren't made to the interface type after it has been
-    // written to file
     pub(crate) fn write_config_file(&self) -> Result<()> {
         let mut cfg_path = Path::new(WICKED_CONFIG_DIR).join(self.name.to_string());
         cfg_path.set_extension(WICKED_FILE_EXT);
@@ -244,14 +229,20 @@ type Result<T> = std::result::Result<T, error::Error>;
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::net_config::NetConfig;
+    use crate::net_config::{self, Interfaces, NetConfigV1};
     use std::path::PathBuf;
     use std::str::FromStr;
 
     fn test_data() -> PathBuf {
-        PathBuf::from(env!("CARGO_MANIFEST_DIR"))
-            .join("test_data")
-            .join("wicked")
+        PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("test_data")
+    }
+
+    fn wicked_config() -> PathBuf {
+        test_data().join("wicked")
+    }
+
+    fn net_config() -> PathBuf {
+        test_data().join("net_config")
     }
 
     // Test the end-to-end trip: "net config from cmdline -> wicked -> serialized XML"
@@ -272,13 +263,13 @@ mod tests {
             "eno8:dhcp4?,dhcp6?",
         ];
         for ok_str in ok {
-            let net_config = NetConfig::from_str(&ok_str).unwrap();
+            let net_config = NetConfigV1::from_str(&ok_str).unwrap();
 
-            for (name, config) in net_config.interfaces {
-                let interface = WickedInterface::from_config(name, config);
+            let wicked_interfaces = net_config.into_wicked_interfaces();
+            for interface in wicked_interfaces {
                 let generated = serde_xml_rs::to_string(&interface).unwrap();
 
-                let mut path = test_data().join(interface.name.to_string());
+                let mut path = wicked_config().join(interface.name.to_string());
                 path.set_extension("xml");
                 let expected = fs::read_to_string(path).unwrap();
 
@@ -290,15 +281,15 @@ mod tests {
     // Test the end to end trip: "net config -> wicked -> serialized XML"
     #[test]
     fn net_config_to_interface_config() {
-        let net_config_str: &str = include_str!("../test_data/net_config/net_config.toml");
-        let net_config: NetConfig = toml::from_str(&net_config_str).unwrap();
+        let net_config_path = net_config().join("net_config.toml");
+        let net_config = net_config::from_path(&net_config_path).unwrap().unwrap();
 
-        for (name, config) in net_config.interfaces {
-            let mut path = test_data().join(&name.to_string());
+        let wicked_interfaces = net_config.into_wicked_interfaces();
+        for interface in wicked_interfaces {
+            let mut path = wicked_config().join(interface.name.to_string());
             path.set_extension("xml");
             let expected = fs::read_to_string(path).unwrap();
 
-            let interface = WickedInterface::from_config(name, config);
             let generated = serde_xml_rs::to_string(&interface).unwrap();
 
             assert_eq!(expected.trim(), generated)
